Split VibesMP into lazy 20-file package
