+++ builds/MicroPython/apps_unfrozen/vibesmp_lib/vibes_library.py
@@ -1,434 +1,9 @@
-# VibesMP path helpers, playlists, scanning, and library database.
-
-# ---- utils.py ----
-
-import os
-
-# Dynamic path resolution to avoid hardcoding /sd/...
-_base_path = ""
-try:
-    _f = __file__
-    if _f.startswith("/"):
-        if "/" in _f:
-            _base_path = _f.rsplit("/", 1)[0] + "/"
-        else:
-            _base_path = "/"
-    else:
-        _cwd = os.getcwd()
-        if not _cwd.endswith("/"): _cwd += "/"
-        if _f.startswith("./"): _f = _f[2:]
-        if "/" in _f:
-            _base_path = _cwd + _f.rsplit("/", 1)[0] + "/"
-        else:
-            _base_path = _cwd
-except (NameError, AttributeError, OSError):
-    # Fallback to standard path if __file__ resolution fails
-    _base_path = "picoware/apps/vibesmp_lib/"
-
-# Ensure _base_path doesn't have /sd/ for Storage API compatibility
-if _base_path.startswith("/sd/"):
-    _base_path = _base_path[4:]
-elif _base_path.startswith("sd/"):
-    _base_path = _base_path[3:]
-
-def get_path(subpath):
-    """Resolve an absolute path within the app package."""
-    path = _base_path + subpath
-    # Final safety check: remove leading / for Storage API
-    if path.startswith("/"):
-        path = path[1:]
-    return path
-
-def format_time(seconds):
-    """Format seconds into MM:SS string."""
-    seconds = int(seconds)
-    return f"{seconds // 60:02}:{seconds % 60:02}"
-
-def get_filename(path):
-    """Get just the filename or folder name from a full path."""
-    if not path:
-        return ""
-    p = path.rstrip("/")
-    res = p.split("/")[-1]
-    if res.lower().endswith(".mp3"):
-        res = res[:-4]
-    return res
-
-def get_parent_path(path):
-    """Get parent directory path with trailing slash."""
-    if not path or path in ("/", "/sd/"):
-        return "/sd/"
-    parts = path.strip("/").split("/")
-    if len(parts) <= 1:
-        return "/sd/"
-    return "/" + "/".join(parts[:-1]) + "/"
-
-def mkdir_p(storage, path):
-    """Create directory and its parents if they don't exist."""
-    if not path or path == "/":
-        return True
-
-    clean_path = path.replace("\\", "/").strip("/")
-    while "//" in clean_path:
-        clean_path = clean_path.replace("//", "/")
-
-    if not clean_path:
-        return True
-
-    if storage.is_directory(clean_path):
-        return True
-
-    parts = clean_path.split("/")
-    for i in range(len(parts)):
-        curr = "/".join(parts[:i+1])
-        if not storage.is_directory(curr):
-            if not storage.mkdir(curr):
-                return False
-    return True
-
-# ---- playlist.py ----
-
-import random
-import json
-
-class Playlist:
-    def __init__(self, storage=None, filename="default.json"):
-        self.storage = storage
-        self.filename = filename
-        self.tracks = []
-        self._current_index = 0
-        self.base_dir = "picoware/vibesmp/playlists/"
-        self.editor_playlist_idx = 0
-        self.editor_library_idx = 0   # Cursor in the library pane of the editor
-        self.active_pane = 0          # 0 = library pane, 1 = playlist pane
-        self._is_dirty = False # Structural dirty (tracks added/removed)
-        self._index_dirty = False # Volatile dirty (index changed)
-        self._save_pending = False
-
-    @property
-    def current_index(self):
-        return self._current_index
-
-    @current_index.setter
-    def current_index(self, value):
-        if self._current_index != value:
-            self._current_index = value
-            self._index_dirty = True
-
-    def __del__(self):
-        self.tracks = []
-        self.storage = None
-
-    def add_track(self, file_path):
-        if file_path:
-            self.tracks.append(file_path)
-            self._is_dirty = True
-
-    def remove_track(self, index):
-        if 0 <= index < len(self.tracks):
-            del self.tracks[index]
-            if index < self.current_index:
-                self.current_index -= 1
-            if len(self.tracks) == 0:
-                self.current_index = 0
-            elif self.current_index >= len(self.tracks):
-                self.current_index = len(self.tracks) - 1
-            self._is_dirty = True
-
-    def clear(self):
-        if self.tracks or self.current_index != 0:
-            self.tracks = []
-            self.current_index = 0
-            self.editor_playlist_idx = 0
-            self._is_dirty = True
-
-    def move_track(self, from_idx, to_idx):
-        if 0 <= from_idx < len(self.tracks) and 0 <= to_idx < len(self.tracks):
-            curr_track = self.get_current()
-            track = self.tracks.pop(from_idx)
-            self.tracks.insert(to_idx, track)
-            self._is_dirty = True
-
-            # Update current_index if it was moved
-            if curr_track:
-                for i, t in enumerate(self.tracks):
-                    if t == curr_track:
-                        self.current_index = i
-                        break
-            return True
-        return False
-
-    def next_track(self, loop_mode, shuffle=False, auto_advance=False):
-        if not self.tracks:
-            return None
-
-        if loop_mode == 1 and auto_advance:  # Loop One
-            return self.tracks[self.current_index]
-
-        if shuffle:
-            self.current_index = random.randint(0, len(self.tracks) - 1)
-        else:
-            if self.current_index + 1 >= len(self.tracks):
-                if loop_mode == 2:  # Loop All
-                    self.current_index = 0
-                else:
-                    # No loop: stay on last track but return None to stop auto-advance
-                    if auto_advance:
-                        return None
-                    else:
-                        # Manual 'next' on last track: stop.
-                        return None
-            else:
-                self.current_index += 1
-
-        return self.tracks[self.current_index]
-
-    def prev_track(self):
-        if not self.tracks:
-            return None
-        self.current_index -= 1
-        if self.current_index < 0:
-            self.current_index = len(self.tracks) - 1
-        return self.tracks[self.current_index]
-
-    def get_current(self):
-        if not self.tracks or self.current_index < 0 or self.current_index >= len(self.tracks):
-            return None
-        return self.tracks[self.current_index]
-
-    def save_as(self, filename, storage_manager=None):
-        """Save current tracks to a new playlist file."""
-        if not filename: return
-        self.filename = filename
-        self._is_dirty = True
-        self.save(force=True, storage_manager=storage_manager)
-
-    def save(self, force=False, storage_manager=None):
-        if not self.storage: return
-        if not self._is_dirty and not force: return
-        from gc import collect
-        collect()
-        try:
-            # Ensure filename doesn't repeat base_dir
-            fname = self.filename
-            if fname.startswith(self.base_dir):
-                fname = fname[len(self.base_dir):]
-
-            full_path = self.base_dir + fname
-            state = {
-                "tracks": self.tracks,
-                "current_index": self.current_index
-            }
-            data = json.dumps(state)
-            if storage_manager:
-                def _mark_saved():
-                    self._save_pending = False
-
-                def _mark_failed():
-                    self._save_pending = False
-                    self._is_dirty = True
-
-                storage_manager.request_write(
-                    full_path,
-                    data,
-                    on_success=_mark_saved,
-                    on_error=_mark_failed,
-                )
-                self._save_pending = True
-                self._is_dirty = False
-            else:
-                if self.storage.write(full_path, data, "w"):
-                    self._save_pending = False
-                    self._is_dirty = False
-                else:
-                    self._save_pending = False
-                    self._is_dirty = True
-            del data
-            collect()
-        except OSError as e:
-            import sys
-            print("[ERROR] playlist.save OSError:", e)
-            sys.print_exception(e)
-            self._save_pending = False
-            self._is_dirty = True
-        except (ValueError, TypeError) as e:
-            import sys
-            print("[ERROR] playlist.save unexpected:", e)
-            sys.print_exception(e)
-            self._save_pending = False
-            self._is_dirty = True
-
-    def load(self, filename=None, storage_manager=None):
-        if self._is_dirty: self.save(storage_manager=storage_manager)
-
-        if filename:
-            self.filename = filename
-        if not self.storage: return
-
-        from gc import collect
-        collect()
-
-        fname = self.filename
-        if fname.startswith(self.base_dir):
-            fname = fname[len(self.base_dir):]
-
-        path = self.base_dir + fname
-        try:
-            data = self.storage.read(path)
-            if data:
-                parsed = json.loads(data)
-                if isinstance(parsed, dict):
-                    self.tracks = parsed.get("tracks", [])
-                    self._current_index = parsed.get("current_index", 0)
-                elif isinstance(parsed, list):
-                    self.tracks = parsed
-                    self._current_index = 0
-
-                # Bounds check
-                if self.current_index >= len(self.tracks):
-                    self.current_index = 0
-
-            self._is_dirty = False
-            del data
-            collect()
-        except (OSError, ValueError) as e:
-            import sys
-            print(f"[ERROR] Playlist: Load failed {path}: {e}")
-            sys.print_exception(e)
-            self.tracks = []
-            self.current_index = 0
-
-# ---- scanner.py ----
-
-def _perf_set(library, name, value):
-    counters = getattr(library, "perf_counters", None)
-    if counters is not None:
-        counters[name] = value
-
-def scan(library, path=None, loading=None, progress_callback=None, quick=False, remove_missing=False):
-    from gc import collect
-    import time
-    collect()
-    start = time.ticks_ms()
-    scan_path = ""
-    print(f"[DEBUG] library: scanning SD root")
-    old_tracks = [library._normalize_track_path(p) for p in getattr(library, "tracks", [])]
-    old_set = set(old_tracks)
-    found = []
-    added = []
-    if not quick:
-        library.tracks = []
-    library._tree_structure = None
-    library._flat_tree_cache = None
-    library._count = 0
-    library._scan_dirs = 0
-    library._scan_last_path = scan_path
-    library._scan_progress_force = False
-
-    try:
-        _recursive_scan(library, scan_path, loading, progress_callback, found, old_set, added, quick)
-        if quick:
-            library.tracks = old_tracks + added
-        elif remove_missing:
-            library.tracks = found
-        if progress_callback:
-            library._scan_progress_force = True
-            try:
-                progress_callback(getattr(library, "_scan_last_path", scan_path), library._count)
-            finally:
-                library._scan_progress_force = False
-        library.save()
-        if hasattr(library, "_sync_state_with_tracks"):
-            library._sync_state_with_tracks(save=True)
-        deferred = 0
-        if hasattr(library, "queue_metadata_for_tracks"):
-            deferred = library.queue_metadata_for_tracks(library.tracks)
-        found_set = set(found)
-        removed = len([p for p in old_tracks if p not in found_set])
-        if remove_missing:
-            before = len(library.tracks)
-            library.tracks = [p for p in library.tracks if p in found_set]
-            if quick:
-                removed = before - len(library.tracks)
-        new_tracks = len([p for p in library.tracks if p not in old_set])
-        unchanged = len([p for p in library.tracks if p in old_set])
-        library.last_scan_summary = {
-            "total": len(library.tracks),
-            "found": len(found),
-            "added": new_tracks,
-            "removed": removed,
-            "unchanged": unchanged,
-            "failed": 0,
-        }
-        _perf_set(library, "scan_ms", time.ticks_diff(time.ticks_ms(), start))
-        _perf_set(library, "scan_tracks", len(library.tracks))
-        _perf_set(library, "scan_dirs", getattr(library, "_scan_dirs", 0))
-        _perf_set(library, "scan_metadata_deferred", deferred)
-        print(f"[DEBUG] library: found {len(library.tracks)} tracks, {new_tracks} new, {removed} removed")
-        return new_tracks if new_tracks > 0 else 0
-    except OSError as e:
-        print(f"[ERROR] library.scan fail:", e)
-        library.last_scan_summary = {
-            "total": len(library.tracks),
-            "found": 0,
-            "added": 0,
-            "removed": 0,
-            "unchanged": 0,
-            "failed": 1,
-        }
-        return 0
-
-def _recursive_scan(library, start_path, loading=None, progress_callback=None, found=None, old_set=None, added=None, quick=False):
-    from gc import collect
-    stack = [start_path]
-    skip_dirs = ("__pycache__", ".git", "System Volume Information", "picoware/apps/vibesmp_lib")
-
-    while stack:
-        path = stack.pop()
-        library._scan_dirs += 1
-        library._scan_last_path = path
-        if progress_callback: progress_callback(path, library._count)
-        try:
-            entries = library.storage.read_directory(path)
-            if not entries:
-                continue
-
-            for entry in entries:
-                if loading: loading.animate()
-                item = entry.get("filename")
-                if not item or item.startswith(".") or item in skip_dirs:
-                    continue
-
-                if not path:
-                    full_path = item
-                else:
-                    full_path = path.rstrip("/") + "/" + item
-
-                if entry.get("is_directory"):
-                    if full_path not in skip_dirs:
-                        stack.append(full_path)
-                elif item.lower().endswith(".mp3"):
-                    library._count += 1
-                    library._scan_last_path = full_path
-                    if progress_callback: progress_callback(full_path, library._count)
-
-                    save_path = "/sd/" + full_path
-                    if found is not None:
-                        found.append(save_path)
-                    if quick:
-                        if save_path not in old_set:
-                            added.append(save_path)
-                    else:
-                        library.tracks.append(save_path)
-
-                    if library._count % 50 == 0: collect()
-        except OSError: pass
-        except ValueError as e:
-            print(f"[ERROR] library.scan unexpected {path}:", e)
+# VibesMP library database.
 
 # ---- vibes_library.py ----
 
 import json
+from vibesmp_lib.utils import mkdir_p
 
 DEBUG_LIBRARY = False
 
@@ -667,7 +242,7 @@ class Library:
             return self._title_cache[path]
         title = ""
         try:
-            from vibesmp_lib.metadata import get_cached_title
+            from vibesmp_lib.metadata_engine import get_cached_title
             title = get_cached_title(self.storage, self, path)
         except (ImportError, OSError, ValueError):
             pass
@@ -715,7 +290,7 @@ class Library:
             return self._meta_cache[path]
         md = {}
         try:
-            from vibesmp_lib.metadata import get_meta_paths
+            from vibesmp_lib.metadata_engine import get_meta_paths
             meta_path, _ = get_meta_paths(self, path)
             n = meta_path
             if n.startswith("/sd/"):
@@ -741,7 +316,7 @@ class Library:
 
     def _metadata_file_exists(self, path):
         try:
-            from vibesmp_lib.metadata import get_meta_paths
+            from vibesmp_lib.metadata_engine import get_meta_paths
             meta_path, _ = get_meta_paths(self, self._normalize_track_path(path))
             if meta_path.startswith("/sd/"):
                 meta_path = meta_path[4:]
@@ -776,7 +351,7 @@ class Library:
                 return True
             start = time.ticks_ms() if self.perf_counters is not None else 0
             try:
-                from vibesmp_lib.metadata import extract_metadata
+                from vibesmp_lib.metadata_engine import extract_metadata
                 ok = extract_metadata(self.storage, path, self)
                 if start:
                     self._perf_timing("metadata_idle_extract", time.ticks_diff(time.ticks_ms(), start))
